neper: use epoll_pwait2 when available
Update neper to use the newer, more granular syscall when available.
There are no behavior changes in this change, as we're still passing
milliseconds into epoll_pwait2. Subsequent PRs utilize the nanosecond
granularity.

Tested that performance is not affected by running on devrez and locally. Local tests looked like the below, with devrez commands being identical but with different addresses and without blaze.

```
$ tcp_rr -H 127.0.0.1 -Q 200 -R 400 -T 400 -F 8000
$ tcp_rr -c -H 127.0.0.1 -Q 200 -R 400 -T 400 -F 8000
```