neper: nanosecond granularity
When running on newer kernels with epoll_pwait2 we schedule with true
nanosecond granularity.

When running on older kernels behavior should be unchanged.

Tested that performance is not affected by running on 2 remotes and
locally. Local tests looked like the below, with remote commands being
identical but with different addresses.

```
$ tcp_rr -H 127.0.0.1 -Q 200 -R 400 -T 400 -F 8000
$ tcp_rr -c -H 127.0.0.1 -Q 200 -R 400 -T 400 -F 8000
```